Added word break/end of string detection to the roller regexs.
I was motivated to make this change when I noticed that dice.Roll was parsing "3d8v3" as a standard roll rather than a Versus roll, with "v3" tagged as the reason. Forcing the std roller to check for either a space or an end-of-string allowed the versus roller to take over in these cases. I went ahead and added the same regex check "($|\s)" to the other patterns as well. This has the advantage of forcing all formats to look the same for a user, as well as ensuring that potentially malformed input, (example: "4b5bl" in the eote roller) would not have the malformed part interpreted as the reason field, but instead return a format error.

There's a small secondary change as well that cropped up as I was messing around with dice.Roll's parsing: the eote regex steps on the versus roller's 'r' option, treating this as a red die rather than a reroll command.